chore: rework due to upstream restructuring
