+++ src/main/java/at/ac/uibk/dps/cirrina/execution/object/action/ActionBuilder.java
@@ -1,16 +1,18 @@
 package at.ac.uibk.dps.cirrina.execution.object.action;
 
+import at.ac.uibk.dps.cirrina.classes.statemachine.StateMachineClassBuilder;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.ActionDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.AssignActionDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.ContextVariableDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.CreateActionDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.EventDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.InvokeActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.TimeoutResetActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.SpawnActionDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.MatchActionDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.MatchCaseDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.RaiseActionDescription;
 import at.ac.uibk.dps.cirrina.csml.description.Csml.TimeoutActionDescription;
-import at.ac.uibk.dps.cirrina.csml.description.Csml.TimeoutResetActionDescription;
 import at.ac.uibk.dps.cirrina.execution.object.context.ContextVariable;
 import at.ac.uibk.dps.cirrina.execution.object.context.ContextVariableBuilder;
 import at.ac.uibk.dps.cirrina.execution.object.event.Event;
